Fix mailbox name initialiser behaviour (#562)
Resolves #560

An invalid UTF-8 string may crash the MailboxName initialiser depending on how the bytes are read into a string. NIO uses String(decoding:, as:) which is not failable, and will mutate bytes to use their unicode equivalents if needed. This isn't what we want in IMAP mailbox names, where byte preservation is important.

The new solution is to actually validate the bytes, and only if they are valid UTF-8 to bother checking if the bytes given represent an "inbox".