Add support for expressing Pipelining Behavior (#569)
Add support for expressing Pipelining Behavior

Motivation:

IMAP supports command pipelining. But there are restrictions as to what commands can be run at the same time.

This is described in RFC 3501 section 5.5. “Multiple Commands in Progress”.

With this change, it’s possible to know if a specific command can be run while a given set of commands is running.