New state machine with boilerplate for IDLE command (#586)
NOTE: The many defaults used when switching over things is just because this is an example. If we're happy with how this looks, I'll remove them and handle properly.

This is a basic demonstrative implementation of the new client state machine that shows how the IDLE command is supported.

For an idle state to be established, the client needs to send the idleStart command, and the server needs to respond with a continuationRequest, at which point the state is idle.