Make Mailbox/Use Attribute casing match the spec (#633)
Resolves #612

The IMAP RFC dictates that casing for constants shouldn't matter, however there was a bug in the constants we were using that made this sometimes untrue, as we used an internal init to provide a fast path for known attributes. This fast path assumed that the constants were already lowercased, however a couple weren't.

The bug has been fixed by removing the fast path. As a nicety towards mail we also now preserve the case, however manually make equality comparisons non-case-sensitive.