Rename modificationSequence to modification (#649)
Resolves #604

Rename modificationSequence to just modified as it might not contain SequenceNumbers. Also change the associated type from SequenceSet (aka MessageIdentifierSet<SequenceNumber>) to UnknownIdentifier, which can then be treated as either UIDs or SequenceNumbers depending on what the user is expecting.

(Note this is required as it is impossible to distinguish between SequenceNumbers and UIDs when parsing).