Enable limiting the number of fetch attributes and body size (#665)
Resolves #402

To prevent a massive amount of fetch attributes from being sent maliciously, the parser now supports a configurable limit on how many fetch attributes are allowed. If this limit is exceeded then an error is thrown.

Added some tests to show that this works for both "simple" attributes, and ones that involve streaming.