Add support for nil encoding to address iCloud issues (#685)
Resolves #684

iCloud is sending nil for message encoding types, which isn't allowed according to the spec. That said, we have to support it regardless, so we now support nil for the message encoding.

Also added some body parsing tests as there apparently weren't any "combined" ones like this. Additionally as there are multiple ways of saying "multipart/mixed" and other "multipart/*" subtypes, we now check and merge them.